Try sending all children in the initial update
Collecting up the list of initial updates on the client side is proving
annoying. This will get harder when we handle permission changes and
other things which cause full updates to be sent at other times.

We are using a WebSocket; there is no particular limit to the length of
a message. There is a small possibility that we will hit some JSON
parser limit but in that case the client probably shouldn't be using
SEARCH in the first place.