Apply ACLs to SEARCH
With the current ACL system a 403 always applies to the parent and all
children. We only check ACLs when we get a child update but then we need
to convert that into a parent update.

The spec says we must not follow a no-access update with a child update;
we need to convert that child update to a full update. We also need to
pretend that we've just sent a no-access update initially.

Implement `withState`, which is the operator I usually want `scan` to
be. This makes a distinction between the state value and the pipeline
output value.